refactor files structure for the data files
