+++ src/main/kotlin/data/weather/WeatherDataSource.kt
@@ -1,0 +1,7 @@
+package org.baghdad.data.weather
+
+import data.source.weather.model.WeatherDto
+
+interface  WeatherDataSource {
+    suspend fun fetchCurrentWeather(city: String): WeatherDto
+}
